requirements.txt
asgiref==3.5.1
astroid==2.11.5
autopep8==1.6.0
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.4
colorclass==2.2.2
dill==0.3.4
Django==4.0.4
docopt==0.6.2
gcloud==0.18.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
isort==5.10.1
jws==0.1.3
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauth2client==4.1.3
packaging==21.3
Pillow==9.1.0
pip-upgrader==1.4.15
platformdirs==2.5.2
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.15.0
pylint==2.13.8
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.20.0
python-jwt==2.0.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.12
wrapt==1.14.1
